requirements.txt
glfw
glm
PyOpenGL 
PyOpenGL_accelerate